Notoj object creation
Notoj annotations are cool but they are "raw". It returns an array (or an object that looks like an Array). It might not be enough, so I'm working on a class loading.

At this stage it will convert annotation to a fully qualified path if it is possible. For instance:

```php
<?php

use Foo\Bar as xxx;

/**
 *  @xxx\yyy
 */
class yyy
{
}
```

The Annotation object has an method `->setFile($file)` which will scan the `$file` reading all classes aliases.

This commit is not completed yet.