Extendend the parser
Allow to call php functions *before* constructing the Annotation object.

`@Foobar(_("some text to be translated"))`

The above statement is valid, if `_` is defined it will be called, if `_` is not found then the first argument will be used instead (if the function was called with one argument, otherwise the array of arguments).